requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.1.0
django-editorjs==0.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-on-heroku==1.1.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.3.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-git==2018.2.1
pytz==2021.1
requests==2.25.1
Send2Trash==1.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
wget==3.2
whitenoise==5.2.0
zipp==3.4.1
